Fixed crash on server
