+++ src/main/java/io/alwa/featuredservers/FeaturedServers.java
@@ -2,14 +2,10 @@ package io.alwa.featuredservers;
 
 import com.google.gson.Gson;
 import com.google.gson.stream.JsonReader;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.multiplayer.ServerData;
-import net.minecraft.client.multiplayer.ServerList;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
-import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -17,17 +13,11 @@ import java.io.File;
 import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 
 @Mod("featuredservers")
 public class FeaturedServers {
-    private static final Logger LOGGER = LogManager.getLogger();
+    public static final Logger LOGGER = LogManager.getLogger();
     private static String FMLConfigFolder;
-    private static ServerList serverList;
-    public static final Map<String, FeaturedServerData> servers = new HashMap<>();
 
     public FeaturedServers() {
         FMLJavaModLoadingContext.get().getModEventBus().addListener(this::onConfigLoad);
@@ -69,60 +59,18 @@ public class FeaturedServers {
         Gson gson = new Gson();
         JsonReader reader = new JsonReader(serversFile);
         ServerDataHelper[] featuredList = gson.fromJson(reader, ServerDataHelper[].class);
-        if (featuredList != null) {
-            serverList = new ServerList(Minecraft.getInstance());
-            for (ServerDataHelper serverhelp : featuredList) {
-                FeaturedServerData server = new FeaturedServerData(serverhelp.serverName, serverhelp.serverIP, false, serverhelp.disableButtons);
-                if(serverhelp.forceResourcePack != null && serverhelp.forceResourcePack) server.setResourcePackStatus(ServerData.ServerResourceMode.ENABLED);
-                if (inList(server, serverList)) {
-                    LOGGER.log(Level.INFO, "Featured server already in server list");
-                } else {
-                    LOGGER.log(Level.INFO, "Adding featured server");
-                    serverList.add(server);
-                    serverList.save();
-                }
-                servers.put(server.ip, server);
-            }
-        }
+        new FeaturedList().doFeaturedListStuff(featuredList);
     }
 
     void onConfigLoad(ModConfig.Loading event) {
         FMLConfigFolder = event.getConfig().getFullPath().getParent().toString();
     }
 
-    public static Boolean inList(ServerData server, ServerList list) {
-        return list != null && toList(list).stream().anyMatch(serverData -> serverData.name != null && serverData.ip != null
-                && serverData.name.equalsIgnoreCase(server.name) && serverData.ip.equalsIgnoreCase(server.ip));
-    }
-
-    public static List<ServerData> toList(ServerList list) {
-        List<ServerData> data = new ArrayList<>();
-        for (int i = 0; i < list.size(); i++) {
-            data.add(list.get(i));
-        }
-        return data;
-    }
-
-    public static ServerList getServerList() {
-        return serverList;
-    }
-
-    public class ServerDataHelper {
+    public static class ServerDataHelper {
 
         public String serverName;
         public String serverIP;
         public Boolean forceResourcePack;
         public Boolean disableButtons;
     }
-
-    public static class FeaturedServerData extends ServerData {
-
-        public final boolean disableButtons;
-
-        public FeaturedServerData(String name, String ip, boolean forceResourcePack, Boolean disableButtons) {
-            super(name, ip, forceResourcePack);
-            this.disableButtons = disableButtons != null && disableButtons;
-        }
-    }
-
 }
